+++ models/order.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
   var Order = sequelize.define('Order', {
     id_user: DataTypes.INTEGER,
     is_checkout: DataTypes.BOOLEAN,
-    total: DataTypes.MONEY
+    total: DataTypes.INTEGER
   }, {
     classMethods: {
       associate: function(models) {
